Add files via upload
Write a program to check whether the given number is a perfect number or not.

A number is considered as a Perfect number if sum of all factors. excluding itself is equal to the number.

Input

The input will be a single line containing a positive integer.

Output

The output should be a single line containing either "Perfect Number" or "Not a Perfect Number".

Explanation

For example, if the given number is 6, the factors of 6 excluding it are 1, 2, and 3. The sum of these factors is 6, which is equal to the given number. So the output should be "Perfect Number.

Similarly, if the given number is 21, the factors of 21 excluding it are 1, 3, and 7. The sum of these factors is 11, Which is not equal to the given number. So the output should be "Not a Perfect Number".